requirements.txt
dash==3.6.0
plotly.express
pandas
